Add a script and Github action to check a PRs age
By convention, we shouldn't adopt an RFC for two weeks (even if it's recieved
enough feedback), to give all team members a chance to wade in. A couple of
PRs have been merged before this time limit though, so this adds a simple
script to check the age of a PR and set a status. If a PR is under two weeks
old, the status is set to pending, if it's over two weeks old, the status is set to
passed. This should be enough to prevent an accidental merge before the
time has elapsed.

The wording is a bit clumsy on the statuses, so I'd welcome some feedback
on a better form of words.