Categorise RFCs
As noted by @thill-dxw (and to borrow his words quite heavily), given we
now have a much wider array of languages and stacks than just Rails and
GovPress, we are likely to start generating a substantial number of RFCs
related to coding standards, templates, frameworks etc

This categorises RFCs, reducing the mental burden of seeing a big list
and having to manually categorise into relevant and irrelevant

With this categorisation, which can develop as appropriate over time, it
will be much easier for new developers to find the RFCs that would be
most beneficial for them to read (i.e. the more general ways of working
plus those relevant to the stack they're working on) in order to be
onboarded on the dxw way of doing things. Developers are more likely to
read a tailored group of RFCs than all of them, which might reduce the
tendency to learn the dxw way by osmosis/code review!

This also changes the casing on `Use existing libraries where possible`
to bring it in line with the rest (sentence case)