Clarify linting RFC
The use of "standard" in the linting RFC has caused some confusion
among developers. While the title is `Use standard` not `Use Standard`
and later it says `meets a standard` not `meet Standard`, when scanning
the RFC it's easy to mistake this for suggesting we use the specific
set of linters called Standard (for languages where they exist, i.e.
JavaScript/TypeScript and Ruby)

This removes use of the word "standard" except where referencing linters
called Standard, and provides examples of other linters that we also use
within the JavaScript/TypeScript domain